Inline `HandlerWrapper` into `Connection`
